backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
asyncpg==0.28.0
sqlalchemy==2.0.20
pydantic==2.3.0
openai==0.28.0
pypdf==3.15.1
python-multipart==0.0.6
python-dotenv==1.0.0
pgvector==0.2.0
numpy==1.25.2
httpx==0.24.1
tiktoken==0.5.1
greenlet==2.0.2